chore: rename mods
