+++ bill/close_bill.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	db "encore.app/internal/db/bill"
-	workflows "encore.app/internal/temporal/bill/workflow"
+	"encore.app/internal/temporal/bill/workflows"
 	"github.com/bojanz/currency"
 	"go.temporal.io/sdk/client"
 )
