+++ bill/increase_bill.go
@@ -8,7 +8,7 @@ import (
 
 	customerrors "encore.app/internal/custom_errors"
 	db "encore.app/internal/db/bill"
-	workflows "encore.app/internal/temporal/bill/workflow"
+	"encore.app/internal/temporal/bill/workflows"
 	"go.temporal.io/sdk/client"
 )
