+++ go.mod
@@ -1,6 +1,6 @@
 module encore.app
 
-go 1.18
+go 1.20
 
 require (
 	encore.dev v1.27.0
