+++ internal/temporal/bill/activities/activity.go
@@ -1,4 +1,4 @@
-package activity
+package activities
 
 import (
 	"context"
